Fix bug when toggling importance of notes
Previosly, we're manually creating a random id before sending a new note
to the json-server. However, this id is a number, and json-server likes
strings. So, we just let json-server take care of assigning random ids
to new notes automatically.

We also convert the ids in the existing db.json into strings.